refactor: Move routes in a specific component
This commit is a preparation for the react-router v6 upgrade
